requirements_test.txt
Django==2.0
djangorestframework==3.7.7
lxml==4.1.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
django-filter==1.1.0
mock==2.0.0